Implement typescript for deny-list.js.
