+++ packages/browser-agent-core/src/common/deny-list/deny-list.interfaces.ts
@@ -1,0 +1,9 @@
+export interface XhrParams {
+  method: string
+  hostname: string
+  port: string
+  protocol: string
+  host: string
+  pathname: string
+  status: number
+}
